basket and some adjustements
