+++ MarketMinds.Shared/ProxyRepository/BasketProxyRepository.cs
@@ -7,10 +7,11 @@ using System.Text.Json.Serialization;
 using System.Threading.Tasks;
 using MarketMinds.Shared.Models;
 using Microsoft.Extensions.Configuration;
+using MarketMinds.Shared.IRepository;
 
 namespace MarketMinds.Shared.ProxyRepository
 {
-    public class BasketProxyRepository
+    public class BasketProxyRepository : IBasketRepository
     {
         public const int MAXIMUM_QUANTITY_PER_ITEM = 10;
         private const int MINIMUM_USER_ID = 0;
@@ -23,14 +24,13 @@ namespace MarketMinds.Shared.ProxyRepository
         private const int INVALID_BASKET_ID = -1;
 
         private readonly HttpClient httpClient;
-        private readonly string apiBaseUrl;
         private readonly JsonSerializerOptions jsonOptions;
 
         // Constructor with configuration
         public BasketProxyRepository(IConfiguration configuration)
         {
             httpClient = new HttpClient();
-            apiBaseUrl = configuration["ApiSettings:BaseUrl"] ?? "http://localhost:5000";
+            var apiBaseUrl = configuration["ApiSettings:BaseUrl"] ?? "http://localhost:5000";
             if (!apiBaseUrl.EndsWith("/"))
             {
                 apiBaseUrl += "/";
@@ -54,502 +54,133 @@ namespace MarketMinds.Shared.ProxyRepository
             httpClient.Timeout = TimeSpan.FromSeconds(30);
         }
 
-        public void AddProductToBasket(int userId, int productId, int quantity)
+        public HttpResponseMessage AddProductToBasketRaw(int userId, int productId, int quantity)
         {
-            if (userId <= MINIMUM_USER_ID)
-            {
-                throw new ArgumentException("Invalid user ID");
-            }
-            if (productId <= MINIMUM_ITEM_ID)
-            {
-                throw new ArgumentException("Invalid product ID");
-            }
-
-            try
-            {
-                // Apply the maximum quantity limit
-                int limitedQuantity = Math.Min(quantity, MAXIMUM_QUANTITY_PER_ITEM);
-
-                // Make the API call to add product to basket
-                var response = httpClient.PostAsJsonAsync($"user/{userId}/product/{productId}", limitedQuantity).Result;
-                response.EnsureSuccessStatusCode();
-            }
-            catch (HttpRequestException ex)
-            {
-                throw new ApplicationException($"Failed to add product to basket: {ex.Message}", ex);
-            }
+            var response = httpClient.PostAsJsonAsync($"user/{userId}/product/{productId}", quantity).Result;
+            response.EnsureSuccessStatusCode();
+            return response;
         }
 
-        public Basket GetBasketByUser(User user)
+        public string GetBasketByUserRaw(int userId)
         {
-            if (user == null || user.Id <= MINIMUM_USER_ID)
-            {
-                throw new ArgumentException("Valid user must be provided");
-            }
-
-            try
-            {
-                string fullUrl = $"{apiBaseUrl}api/basket/user/{user.Id}";
-
-                var response = httpClient.GetAsync(fullUrl).Result;
-                response.EnsureSuccessStatusCode();
-
-                // Use the custom JSON converter to properly deserialize the response
-                var responseContent = response.Content.ReadAsStringAsync().Result;
-
-                try
-                {
-                    // Create improved JsonSerializerOptions optimized for the client
-                    var options = new JsonSerializerOptions
-                    {
-                        PropertyNameCaseInsensitive = true,
-                        AllowTrailingCommas = true,
-                        ReadCommentHandling = JsonCommentHandling.Skip,
-                        NumberHandling = JsonNumberHandling.AllowReadingFromString,
-                        PropertyNamingPolicy = JsonNamingPolicy.CamelCase,
-                        // Use IgnoreCycles instead of Preserve
-                        ReferenceHandler = ReferenceHandler.IgnoreCycles
-                    };
-
-                    // Add UserJsonConverter to handle password field type mismatch
-                    options.Converters.Add(new MarketMinds.Shared.Services.UserJsonConverter());
-
-                    // Deserialize directly
-                    var basket = JsonSerializer.Deserialize<Basket>(responseContent, options);
-
-                    // Make sure we have an Items collection
-                    if (basket.Items == null)
-                    {
-                        basket.Items = new List<BasketItem>();
-                    }
-
-                    return basket;
-                }
-                catch (JsonException ex)
-                {
-                    // Try to fallback to a simpler deserialization
-                    var fallbackBasket = new Basket { Id = user.Id, Items = new List<BasketItem>() };
-                    return fallbackBasket;
-                }
-            }
-            catch (HttpRequestException ex)
-            {
-                throw new ApplicationException($"Failed to retrieve user's basket: {ex.Message}", ex);
-            }
-            catch (Exception ex)
-            {
-                throw new ApplicationException("Failed to retrieve user's basket", ex);
-            }
+            string fullUrl = $"{httpClient.BaseAddress.AbsoluteUri.Replace("api/basket/", "")}api/basket/user/{userId}";
+            var response = httpClient.GetAsync(fullUrl).Result;
+            response.EnsureSuccessStatusCode();
+            return response.Content.ReadAsStringAsync().Result;
         }
 
-        public void RemoveProductFromBasket(int userId, int productId)
+        public HttpResponseMessage RemoveProductFromBasketRaw(int userId, int productId)
         {
-            if (userId <= MINIMUM_USER_ID)
-            {
-                throw new ArgumentException("Invalid user ID");
-            }
-            if (productId <= MINIMUM_ITEM_ID)
-            {
-                throw new ArgumentException("Invalid product ID");
-            }
-
-            try
-            {
-                // Make the API call to remove product from basket
-                var response = httpClient.DeleteAsync($"user/{userId}/product/{productId}").Result;
-                response.EnsureSuccessStatusCode();
-            }
-            catch (HttpRequestException ex)
-            {
-                throw new ApplicationException($"Failed to remove product from basket: {ex.Message}", ex);
-            }
-            catch (Exception ex)
-            {
-                throw new ApplicationException($"Failed to remove product from basket: {ex.Message}", ex);
-            }
+            var response = httpClient.DeleteAsync($"user/{userId}/product/{productId}").Result;
+            response.EnsureSuccessStatusCode();
+            return response;
         }
 
-        public void UpdateProductQuantity(int userId, int productId, int quantity)
+        public HttpResponseMessage UpdateProductQuantityRaw(int userId, int productId, int quantity)
         {
-            if (userId <= MINIMUM_USER_ID)
-            {
-                throw new ArgumentException("Invalid user ID");
-            }
-            if (productId <= MINIMUM_ITEM_ID)
-            {
-                throw new ArgumentException("Invalid product ID");
-            }
-            if (quantity < MINIMUM_QUANTITY)
-            {
-                throw new ArgumentException("Quantity cannot be negative");
-            }
-
-            try
-            {
-                // Apply the maximum quantity limit
-                int limitedQuantity = Math.Min(quantity, MAXIMUM_QUANTITY_PER_ITEM);
-
-                // Make the API call to update product quantity
-                var response = httpClient.PutAsJsonAsync($"user/{userId}/product/{productId}", limitedQuantity).Result;
-                response.EnsureSuccessStatusCode();
-            }
-            catch (HttpRequestException ex)
-            {
-                throw new ApplicationException($"Failed to update product quantity: {ex.Message}", ex);
-            }
-            catch (Exception ex)
-            {
-                throw new ApplicationException($"Failed to update product quantity: {ex.Message}", ex);
-            }
+            var response = httpClient.PutAsJsonAsync($"user/{userId}/product/{productId}", quantity).Result;
+            response.EnsureSuccessStatusCode();
+            return response;
         }
 
-        public bool ValidateQuantityInput(string quantityText, out int quantity)
+        public HttpResponseMessage ClearBasketRaw(int userId)
         {
-            // Initialize output parameter
-            quantity = MINIMUM_QUANTITY;
-
-            // Check if the input is empty
-            if (string.IsNullOrWhiteSpace(quantityText))
-            {
-                return false;
-            }
-
-            // Try to parse the input string to an integer
-            if (!int.TryParse(quantityText, out quantity))
-            {
-                return false;
-            }
-
-            // Check if quantity is non-negative
-            if (quantity < MINIMUM_QUANTITY)
-            {
-                return false;
-            }
-
-            return true;
+            var response = httpClient.DeleteAsync($"user/{userId}/clear").Result;
+            response.EnsureSuccessStatusCode();
+            return response;
         }
 
-        public int GetLimitedQuantity(int quantity)
+        public string ValidateBasketBeforeCheckOutRaw(int basketId)
         {
-            return Math.Min(quantity, MAXIMUM_QUANTITY_PER_ITEM);
+            var response = httpClient.GetAsync($"{basketId}/validate").Result;
+            response.EnsureSuccessStatusCode();
+            return response.Content.ReadAsStringAsync().Result;
         }
 
-        public void ClearBasket(int userId)
+        public HttpResponseMessage ApplyPromoCodeRaw(int basketId, string code)
         {
-            if (userId <= MINIMUM_USER_ID)
-            {
-                throw new ArgumentException("Invalid user ID");
-            }
-
-            try
-            {
-                // Clear the basket through API
-                var response = httpClient.DeleteAsync($"user/{userId}/clear").Result;
-                response.EnsureSuccessStatusCode();
-            }
-            catch (HttpRequestException ex)
-            {
-                throw new InvalidOperationException($"Could not clear basket: {ex.Message}", ex);
-            }
-            catch (Exception ex)
-            {
-                throw new InvalidOperationException($"Could not clear basket: {ex.Message}", ex);
-            }
+            var response = httpClient.PostAsJsonAsync($"{basketId}/promocode", code).Result;
+            return response;
         }
 
-        public bool ValidateBasketBeforeCheckOut(int basketId)
+        public string GetPromoCodeDiscountRaw(string code)
         {
-            if (basketId <= MINIMUM_BASKET_ID)
+            string normalizedCode = code.Trim().ToUpper();
+            var response = httpClient.PostAsJsonAsync($"1/promocode", normalizedCode).Result;
+
+            if (response.IsSuccessStatusCode)
             {
-                throw new ArgumentException("Invalid basket ID");
+                return response.Content.ReadAsStringAsync().Result;
             }
 
-            try
-            {
-                // Validate basket through API
-                var response = httpClient.GetAsync($"{basketId}/validate").Result;
-                response.EnsureSuccessStatusCode();
-
-                var responseContent = response.Content.ReadAsStringAsync().Result;
-
-                try
-                {
-                    return JsonSerializer.Deserialize<bool>(responseContent, jsonOptions);
-                }
-                catch (JsonException)
-                {
-                    // Try a simple parsing approach as fallback
-                    responseContent = responseContent.Trim().ToLower();
-                    if (responseContent == "true")
-                    {
-                        return true;
-                    }
-
-                    return false;
-                }
-            }
-            catch (HttpRequestException ex)
-            {
-                throw new InvalidOperationException($"Could not validate basket: {ex.Message}", ex);
-            }
-            catch (Exception ex)
-            {
-                throw new InvalidOperationException($"Could not validate basket: {ex.Message}", ex);
-            }
+            return "{}";
         }
 
-        public void ApplyPromoCode(int basketId, string code)
+        public string CalculateBasketTotalsRaw(int basketId, string promoCode)
         {
-            if (basketId <= MINIMUM_BASKET_ID)
+            string endpoint = $"{basketId}/totals";
+            if (!string.IsNullOrEmpty(promoCode))
             {
-                throw new ArgumentException("Invalid basket ID");
-            }
-            if (string.IsNullOrWhiteSpace(code))
-            {
-                throw new ArgumentException("Promo code cannot be empty");
+                endpoint += $"?promoCode={promoCode}";
             }
 
-            try
-            {
-                // Apply promo code through API
-                var response = httpClient.PostAsJsonAsync($"{basketId}/promocode", code).Result;
-
-                if (!response.IsSuccessStatusCode)
-                {
-                    throw new InvalidOperationException("Invalid promo code");
-                }
-            }
-            catch (HttpRequestException ex)
-            {
-            }
-            catch (Exception ex)
-            {
-                throw ex;
-            }
+            var response = httpClient.GetAsync(endpoint).Result;
+            response.EnsureSuccessStatusCode();
+            return response.Content.ReadAsStringAsync().Result;
         }
 
-        // Class for deserializing discount rate response
-        private class DiscountResponse
+        public HttpResponseMessage DecreaseProductQuantityRaw(int userId, int productId)
         {
-            public double DiscountRate { get; set; }
+            var response = httpClient.PutAsync($"user/{userId}/product/{productId}/decrease", null).Result;
+            response.EnsureSuccessStatusCode();
+            return response;
         }
 
-        // Add a new method to get the discount for a promo code
-        public double GetPromoCodeDiscount(string code, double subtotal)
+        public HttpResponseMessage IncreaseProductQuantityRaw(int userId, int productId)
         {
-            if (string.IsNullOrWhiteSpace(code))
-            {
-                return MINIMUM_DISCOUNT;
-            }
-
-            try
-            {
-                string normalizedCode = code.Trim().ToUpper();
-                var response = httpClient.PostAsJsonAsync($"1/promocode", normalizedCode).Result;
-
-                if (response.IsSuccessStatusCode)
-                {
-                    var responseContent = response.Content.ReadAsStringAsync().Result;
-                    var result = JsonSerializer.Deserialize<DiscountResponse>(responseContent, jsonOptions);
-                    double discount = subtotal * result.DiscountRate;
-                    return discount;
-                }
-
-                return MINIMUM_DISCOUNT;
-            }
-            catch (Exception ex)
-            {
-                return MINIMUM_DISCOUNT;
-            }
+            var response = httpClient.PutAsync($"user/{userId}/product/{productId}/increase", null).Result;
+            response.EnsureSuccessStatusCode();
+            return response;
         }
 
-        // Add a new method to calculate basket totals
-        public BasketTotals CalculateBasketTotals(int basketId, string promoCode)
+        public async Task<HttpResponseMessage> CheckoutBasketRaw(int userId, int basketId, object requestData)
         {
-            if (basketId <= MINIMUM_BASKET_ID)
-            {
-                throw new ArgumentException("Invalid basket ID");
-            }
+            using var httpClientAccount = new HttpClient();
+            httpClientAccount.BaseAddress = new Uri(httpClient.BaseAddress.AbsoluteUri.Replace("api/basket/", ""));
+            httpClientAccount.Timeout = TimeSpan.FromSeconds(30);
 
-            try
-            {
-                // Get totals from API
-                string endpoint = $"{basketId}/totals";
-                if (!string.IsNullOrEmpty(promoCode))
-                {
-                    endpoint += $"?promoCode={promoCode}";
-                }
-
-                var response = httpClient.GetAsync(endpoint).Result;
-                response.EnsureSuccessStatusCode();
-
-                var responseContent = response.Content.ReadAsStringAsync().Result;
-
-                try
-                {
-                    var totals = JsonSerializer.Deserialize<BasketTotals>(responseContent, jsonOptions);
-                    return totals;
-                }
-                catch (JsonException)
-                {
-                    // Create a default totals object
-                    return new BasketTotals();
-                }
-            }
-            catch (HttpRequestException ex)
-            {
-                throw new InvalidOperationException($"Could not calculate basket totals: {ex.Message}", ex);
-            }
-            catch (Exception ex)
-            {
-                throw new InvalidOperationException($"Could not calculate basket totals: {ex.Message}", ex);
-            }
+            return await httpClientAccount.PostAsJsonAsync($"api/account/{userId}/orders", requestData);
         }
 
-        public void DecreaseProductQuantity(int userId, int productId)
+        // Legacy methods that throw NotImplementedException
+        public void AddItemToBasket(int basketId, int productId, int quantity)
         {
-            if (userId <= MINIMUM_USER_ID)
-            {
-                throw new ArgumentException("Invalid user ID");
-            }
-            if (productId <= MINIMUM_ITEM_ID)
-            {
-                throw new ArgumentException("Invalid product ID");
-            }
-
-            try
-            {
-                // Decrease quantity through API
-                var response = httpClient.PutAsync($"user/{userId}/product/{productId}/decrease", null).Result;
-                response.EnsureSuccessStatusCode();
-            }
-            catch (HttpRequestException ex)
-            {
-                throw new InvalidOperationException($"Could not decrease quantity: {ex.Message}", ex);
-            }
-            catch (Exception ex)
-            {
-                throw new InvalidOperationException($"Could not decrease quantity: {ex.Message}", ex);
-            }
+            throw new NotImplementedException("This method should be called from the service layer");
         }
 
-        public void IncreaseProductQuantity(int userId, int productId)
+        public void ClearBasket(int basketId)
         {
-            if (userId <= MINIMUM_USER_ID)
-            {
-                throw new ArgumentException("Invalid user ID");
-            }
-            if (productId <= MINIMUM_ITEM_ID)
-            {
-                throw new ArgumentException("Invalid product ID");
-            }
-
-            try
-            {
-                // Increase quantity through API
-                var response = httpClient.PutAsync($"user/{userId}/product/{productId}/increase", null).Result;
-                response.EnsureSuccessStatusCode();
-            }
-            catch (HttpRequestException ex)
-            {
-                throw new InvalidOperationException($"Could not increase quantity: {ex.Message}", ex);
-            }
-            catch (Exception ex)
-            {
-                throw new InvalidOperationException($"Could not increase quantity: {ex.Message}", ex);
-            }
+            throw new NotImplementedException("This method should be called from the service layer");
         }
 
-        public async Task<bool> CheckoutBasketAsync(int userId, int basketId, double discountAmount = 0, double totalAmount = 0)
+        public List<BasketItem> GetBasketItems(int basketId)
         {
-            if (userId <= INVALID_USER_ID)
-            {
-                throw new ArgumentException("Invalid user ID");
-            }
+            throw new NotImplementedException("This method should be called from the service layer");
+        }
 
-            if (basketId <= INVALID_BASKET_ID)
-            {
-                throw new ArgumentException("Invalid basket ID");
-            }
+        public Basket GetBasketByUserId(int userId)
+        {
+            throw new NotImplementedException("This method should be called from the service layer");
+        }
 
-            try
-            {
-                // First, validate the basket before checkout
-                if (!ValidateBasketBeforeCheckOut(basketId))
-                {
-                    throw new InvalidOperationException("Basket validation failed");
-                }
+        public void RemoveItemByProductId(int basketId, int productId)
+        {
+            throw new NotImplementedException("This method should be called from the service layer");
+        }
 
-                // Get the current basket totals
-                var basketTotals = CalculateBasketTotals(basketId, null);
-                System.Diagnostics.Debug.WriteLine($"DIAGNOSTIC: CheckoutBasketAsync - Raw basketTotal: {basketTotals.TotalAmount}, Provided discount: {discountAmount}, Provided total: {totalAmount}");
-
-                // Use provided values if they are valid
-                if (discountAmount > 0 && totalAmount > 0)
-                {
-                    System.Diagnostics.Debug.WriteLine($"DIAGNOSTIC: Using provided discount ({discountAmount}) and total ({totalAmount})");
-                    basketTotals.Discount = discountAmount;
-                    basketTotals.TotalAmount = totalAmount;
-                }
-                else
-                {
-                    System.Diagnostics.Debug.WriteLine($"DIAGNOSTIC: No valid discount/total provided, using calculated totals");
-                }
-
-                // Create an HTTP client for account API
-                using var httpClientAccount = new HttpClient();
-                httpClientAccount.BaseAddress = new Uri(apiBaseUrl);
-                httpClientAccount.Timeout = TimeSpan.FromSeconds(30);
-
-                // Create request for the account API to create orders from the basket
-                // Include discount information
-                var request = new
-                {
-                    BasketId = basketId,
-                    DiscountAmount = basketTotals.Discount,
-                    TotalAmount = basketTotals.TotalAmount
-                };
-
-                System.Diagnostics.Debug.WriteLine($"Checkout request: BasketId={basketId}, Discount={basketTotals.Discount}, Total={basketTotals.TotalAmount}");
-
-                // Call the account API to create the order
-                var response = await httpClientAccount.PostAsJsonAsync($"api/account/{userId}/orders", request);
-
-                // Handle the response
-                if (response.IsSuccessStatusCode)
-                {
-                    // If successful order creation, the API should have cleared the basket already
-                    System.Diagnostics.Debug.WriteLine($"Successfully created order from basket {basketId} for user {userId}");
-                    return true;
-                }
-                else
-                {
-                    var errorContent = await response.Content.ReadAsStringAsync();
-                    System.Diagnostics.Debug.WriteLine($"Error creating order: {response.StatusCode}, Content: {errorContent}");
-
-                    // If it's a balance issue (400 status code), throw a more user-friendly message
-                    if (response.StatusCode == System.Net.HttpStatusCode.BadRequest &&
-                        errorContent.Contains("Insufficient funds"))
-                    {
-                        throw new InvalidOperationException(errorContent);
-                    }
-
-                    return false;
-                }
-            }
-            catch (HttpRequestException ex)
-            {
-                throw new InvalidOperationException($"Could not checkout basket: {ex.Message}", ex);
-            }
-            catch (InvalidOperationException)
-            {
-                // Re-throw InvalidOperationException as-is to preserve the message
-                throw;
-            }
-            catch (Exception ex)
-            {
-                throw new InvalidOperationException($"Could not checkout basket: {ex.Message}", ex);
-            }
+        public void UpdateItemQuantityByProductId(int basketId, int productId, int quantity)
+        {
+            throw new NotImplementedException("This method should be called from the service layer");
         }
     }
 }
